Add release script snapshot test

+++ scripts/release/prepare-stable-commands/check-out-packages.js
@@ -3,11 +3,21 @@
 'use strict';
 
 const {exec} = require('child-process-promise');
+const {existsSync} = require('fs');
 const {join} = require('path');
 const {logPromise} = require('../utils');
 const theme = require('../theme');
 
-const run = async ({cwd, packages, version}) => {
+const run = async ({cwd, local, packages, version}) => {
+  if (local) {
+    // Sanity test
+    if (!existsSync(join(cwd, 'build', 'node_modules', 'react'))) {
+      console.error(theme.error`No local build exists.`);
+      process.exit(1);
+    }
+    return;
+  }
+
   // Cleanup from previous builds
   await exec(`rm -rf ./build/node_modules*`, {cwd});
   await exec(`mkdir ./build/node_modules`, {cwd});
